implements protocol to simplify rederer for overlay method
