+++ PhotoPoints/Model/OverlayData.swift
